+++ agent/algorithm/wasm/wasm.go
@@ -59,6 +59,10 @@ func (w *wasm) Stop() error {
 		return nil
 	}
 
+	if w.cmd.ProcessState != nil && w.cmd.ProcessState.Exited() {
+		return nil
+	}
+
 	if err := w.cmd.Process.Kill(); err != nil {
 		return fmt.Errorf("error stopping algorithm: %v", err)
 	}
